Exceptions are moved to their own package
