--- src/main/java/com/osrm/client/DistanceService.java
@@ -1,7 +1,0 @@
-package com.osrm.client;
-
-import java.util.List;
-
-public interface DistanceService {
-    DistanceMatrix buildDistanceMatrix(List<GeoLocation> coordinates, double speedRate, String country, String token, String profile, String options);
-}
