+++ src/main/java/com/osrm/client/exception/DistanceMatrixResponseException.java
@@ -1,4 +1,4 @@
-package com.osrm.client;
+package com.osrm.client.exception;
 
 public class DistanceMatrixResponseException extends RuntimeException {
   public DistanceMatrixResponseException(String message) {
