+++ src/main/java/com/osrm/client/exception/OptimizationException.java
@@ -1,4 +1,4 @@
-package com.osrm.client;
+package com.osrm.client.exception;
 
 public abstract class OptimizationException extends RuntimeException {
     public OptimizationException(String message) {
